chore: rename api routes
