+++ src/app/api/suscription/route.ts
@@ -1,7 +1,7 @@
-import { ResourceSchema, Suscriptor, SuscriptorSchema } from "@/app/schema"
+import {  Suscriptor, SuscriptorSchema } from "@/app/schema"
 import { DataSource } from "@/services/datasource"
 import { NextResponse, NextRequest } from "next/server"
-import { z, ZodError } from "zod"
+import {  ZodError } from "zod"
 //import { createPreference } from "@/services/mercadopago"
 
 export async function POST(req: NextRequest) { 
@@ -13,10 +13,14 @@ export async function POST(req: NextRequest) {
       ...data,
       resource_id: body.resource_id,
     })
+
+    if(response.status !== 201) {
+      return NextResponse.json({ error: 'No se pudo completar el registro. Intentá más tarde.' }, { status: 500 })
+    }
     
     //TODO: Launch welcome email with payment_mathods
 
-    return NextResponse.json({ message: 'Inscripción exitosa', data: response })
+    return NextResponse.json({ message: 'Inscripción exitosa', data: response, redirect_url: `/congrats?resource_id=${body.resource_id}` }, { status: 201 })
   } catch (error: unknown) {
     if (error instanceof ZodError) {
       return NextResponse.json(
@@ -31,19 +35,5 @@ export async function POST(req: NextRequest) {
       { error: `No se pudo completar el registro. Intentá más tarde. Error: ${(error as Error).message}` },
       { status: 500 }
     )
-  }
-}
-
-export async function GET(req: NextRequest) {
-  const resourceId = req.nextUrl.searchParams.get('resource_id')
-  if(!resourceId) {
-    return NextResponse.json({ error: 'El id del recurso es requerido' }, { status: 400 })
-  }
-  try { 
-    const response: z.infer<typeof ResourceSchema>[] = await new DataSource().getResource(resourceId)
-    return NextResponse.json({ message: `Recurso ${response[0].name} obtenido`, data: response })
-  } catch (error) {
-    console.error('Error en /api/register:', (error as Error).message)
-    return NextResponse.json({ error: `No se pudo obtener el recurso. Intentá más tarde. Error: ${(error as Error).message}` }, { status: 500 })
   }
 }
